requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
numpy>=1.20.0
pandas>=1.5.0
scikit-learn>=1.2.0
requests>=2.28.0
python-multipart>=0.0.5
google-cloud-bigquery>=3.0.0
googlemaps>=4.0.0
python-dotenv>=0.19.0
aiofiles>=0.8.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
